Restructuring javascript code hints, As we need scopemanager and tern node domain for javascript code refactoring. To resue this code we are restructuring this code so we can keep both extension seperate
