+++ src/JSUtils/HintUtils.js
