+++ src/JSUtils/node/ExtractFileContent.js
