+++ src/brackets.js
@@ -150,6 +150,12 @@ define(function (require, exports, module) {
     require("search/FindInFilesUI");
     require("search/FindReplace");
 
+    //Load common JS module
+    require("JSUtils/HintUtils");
+    require("JSUtils/MessageIds");
+    require("JSUtils/Preferences");
+    require("JSUtils/ScopeManager");
+
     // Compatibility shim for PanelManager to WorkspaceManager migration
     require("view/PanelManager");
